Ignore writable overlaybd commit for accel-layer.
According to containerd's workflow, the snapshotter.Commit() is
invoked during Pull() with Unpack, in order to prepare and commit a
snapshot.

If the snapshotter works in "rwDev", it will commit the writable
overlaybd during the Commit(), which is not required for accel-layer.
Since accel-layer return an overlay mount and containerd has already
apply the layer. In this case, we should ignore the writable overlaybd
commit for accel-layer.

Signed-off-by: Xinran Wang <xin-ran.wang@intel.com>