go.mod: update containerd dependency to stable v2.0.x
Updating to the v2.0.4 patch release as minimum, which includes
[c8d@3d53430] to reduce some (indirect) dependencies, and [c8d@f341477]
to downgrade some unreleased versions of dependencies. The v2.0.4
release included a security fix for [CVE-2024-40635], so updating to
that version to avoid potential security alerts. I did not update to
the latest version because this project is used as dependency for
other modules, and containerd v2.0.5 also updated the minimum Go version
to 1.23.

[c8d@3d53430]: https://github.com/containerd/containerd/commit/3d53430fe14eb76849a6c997d60b21a9f95c19ed
[c8d@f341477]: https://github.com/containerd/containerd/commit/f34147772bb97ef3220c85730b6139bfbf369095
[CVE-2024-40635]: https://github.com/containerd/containerd/security/advisories/GHSA-265r-hfxg-fhmg

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>